PROM_MODEL/requirements.txt
absl-py==0.9.0
appdirs==1.4.4
appnope==0.1.0
astor==0.8.1
backcall==0.2.0
cffi==1.14.0
cmdstanpy==0.4.0
convertdate==2.2.0
cycler==0.10.0
Cython==0.29.15
DBUtils==1.3
decorator==4.4.2
distlib==0.3.0
ephem==3.7.7.1
fbprophet==0.6
filelock==3.0.12
flake8==3.8.3
gast==0.3.3
grpcio==1.27.2
h5py==2.10.0
holidays==0.10.1
importlib-metadata==1.6.0
importlib-resources==1.5.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lightgbm==2.2.3
LunarCalendar==0.0.9
Markdown==3.2.1
matplotlib==3.2.0
mccabe==0.6.1
mock==4.0.1
numpy==1.16.4
pandas==0.24.2
parso==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
plotly==4.5.2
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
py4j==0.10.9
pycodestyle==2.6.0
pycparser==2.19
pyflakes==2.2.0
Pygments==2.6.1
PyMeeus==0.3.7
pyparsing==2.4.6
pyramid-arima==0.8.1
pyspark==3.0.0
pystan==2.19.1.1
python-dateutil==2.8.1
pytz==2019.3
retrying==1.3.3
scikit-learn==0.21.3
scipy==1.0.0
setuptools-git==1.2
six==1.14.0
statsmodels==0.10.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-probability==0.6.0
termcolor==1.1.0
tornado==6.0.3
traitlets==4.3.3
virtualenv==20.0.20
wcwidth==0.2.5
Werkzeug==1.0.0
xgboost==0.90
zipp==3.1.0
